travel1_chatbot/backend/requirements.txt
fastapi
uvicorn
qdrant-client
httpx
python-dotenv
pydantic-settings
python-multipart
pyMuPDF 
python-docx
google-generativeai
requests
sentence-transformers
numpy
pdfplumber
